requirements.txt
geopandas
ipykernel
streamlit
overpy
geopy
streamlit-folium